# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blue Mogul Enterprise Bookkeeping AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /taxes:
    get:
      operationId: listTaxItems
      tags: [taxes]
      summary: List all tax items
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: quarter
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [sales_tax, estimated_quarterly, payroll, deduction]
      responses:
        "200":
          description: List of tax items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxItem"
    post:
      operationId: createTaxItem
      tags: [taxes]
      summary: Create a new tax item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaxItemBody"
      responses:
        "201":
          description: Created tax item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxItem"
  /taxes/{id}:
    put:
      operationId: updateTaxItem
      tags: [taxes]
      summary: Update a tax item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaxItemBody"
      responses:
        "200":
          description: Updated tax item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxItem"
    delete:
      operationId: deleteTaxItem
      tags: [taxes]
      summary: Delete a tax item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /taxes/summary:
    get:
      operationId: getTaxSummary
      tags: [taxes]
      summary: Get tax summary for year-end preparation
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tax summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxSummary"
  /reports/cashflow:
    get:
      operationId: getCashflowReport
      tags: [reports]
      summary: Get cash flow report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Cash flow report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashflowReport"
  /reports/profit-loss:
    get:
      operationId: getProfitLossReport
      tags: [reports]
      summary: Get profit and loss report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: P&L report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitLossReport"
...
components:
  schemas:
...
    TaxItem:
      type: object
      properties:
        id:
          type: integer
        year:
          type: integer
        quarter:
          type: integer
          nullable: true
        type:
          type: string
          enum: [sales_tax, estimated_quarterly, payroll, deduction]
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          format: date
          nullable: true
        paid:
          type: boolean
        paidDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - year
        - type
        - description
        - amount
        - paid
        - createdAt
    CreateTaxItemBody:
      type: object
      properties:
        year:
          type: integer
        quarter:
          type: integer
        type:
          type: string
          enum: [sales_tax, estimated_quarterly, payroll, deduction]
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          format: date
        notes:
          type: string
      required:
        - year
        - type
        - description
        - amount
    UpdateTaxItemBody:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
          format: date
        paid:
          type: boolean
        paidDate:
          type: string
          format: date
        notes:
          type: string
    TaxSummary:
      type: object
      properties:
        year:
          type: integer
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        netIncome:
          type: number
        totalDeductions:
          type: number
        taxableIncome:
          type: number
        salesTaxCollected:
          type: number
        estimatedTaxPaid:
          type: number
        quarterBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/QuarterTaxBreakdown"
        topDeductions:
          type: array
          items:
            $ref: "#/components/schemas/DeductionItem"
      required:
        - year
        - totalRevenue
        - totalExpenses
        - netIncome
        - totalDeductions
        - taxableIncome
        - salesTaxCollected
        - estimatedTaxPaid
        - quarterBreakdown
        - topDeductions
    QuarterTaxBreakdown:
      type: object
      properties:
        quarter:
          type: integer
        revenue:
          type: number
        expenses:
          type: number
        netIncome:
          type: number
        estimatedTaxDue:
          type: number
      required:
        - quarter
        - revenue
        - expenses
        - netIncome
        - estimatedTaxDue
    DeductionItem:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
      required:
        - category
        - amount
    CashflowReport:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        totalInflow:
          type: number
        totalOutflow:
          type: number
        netCashflow:
          type: number
        openingBalance:
          type: number
        closingBalance:
          type: number
        byMonth:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyFlow"
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryFlow"
      required:
        - startDate
        - endDate
        - totalInflow
        - totalOutflow
        - netCashflow
        - openingBalance
        - closingBalance
        - byMonth
        - byCategory
    MonthlyFlow:
      type: object
      properties:
        month:
          type: string
        inflow:
          type: number
        outflow:
          type: number
        net:
          type: number
      required:
        - month
        - inflow
        - outflow
        - net
    CategoryFlow:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
      required:
        - category
        - amount
        - type
    ProfitLossReport:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        grossRevenue:
          type: number
        costOfGoodsSold:
          type: number
        grossProfit:
          type: number
        totalExpenses:
          type: number
        netIncome:
          type: number
        incomeByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryFlow"
        expensesByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryFlow"
      required:
        - startDate
        - endDate
        - grossRevenue
        - costOfGoodsSold
        - grossProfit
        - totalExpenses
        - netIncome
        - incomeByCategory
        - expensesByCategory
...