# lib/api-zod/src/generated/types/deductionItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blue Mogul Enterprise Bookkeeping AI API
 * OpenAPI spec version: 0.1.0
 */

export interface DeductionItem {
  category: string;
  amount: number;
}