# lib/api-zod/src/generated/types/updateTaxItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blue Mogul Enterprise Bookkeeping AI API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTaxItemBody {
  description?: string;
  amount?: number;
  dueDate?: Date;
  paid?: boolean;
  paidDate?: Date;
  notes?: string;
}